Extract Error Boundary into separate file
