+++ frontend/app/analytics-script.tsx
@@ -1,0 +1,15 @@
+export function AnalyticsScript({
+  analyticsDomain,
+}: {
+  analyticsDomain: string | undefined;
+}) {
+  return (
+    !!analyticsDomain && (
+      <script
+        defer
+        data-domain={analyticsDomain}
+        src="https://plausible.io/js/script.js"
+      ></script>
+    )
+  );
+}
